Import, plot and analyze bathymetric and topographic data

marmap2 is an experimental modernization of marmap, a package designed for downloading, plotting and manipulating bathymetric and topographic data in R. It can query the ETOPO 2022 bathymetry and topography database hosted by the NOAA, use simple latitude-longitude-depth data in ascii format, and take advantage of the advanced plotting tools available in R to build publication-quality bathymetric maps. Functions to query data (bathymetry, sampling information, etc...) are available interactively by clicking on marmap maps. Bathymetric and topographic data can also be used to calculate projected surface areas within specified depth/altitude intervals, and constrain the calculation of realistic shortest path distances.

Details

Package:	marmap2

Type:	Package

Version:	0.0.0.9000

Import, plot and analyze bathymetric and topographic data

References

Pante E, Simon-Bouhet B (2013) marmap: A Package for Importing, Plotting and Analyzing Bathymetric and Topographic Data in R. PLoS ONE 8(9): e73051. doi:10.1371/journal.pone.0073051

See also

Author

Eric Pante, Benoit Simon-Bouhet and Jean-Olivier Irisson

Maintainer: Benoit Simon-Bouhet besibo@gmail.com